+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - LitVar2 links: variant link for variants with dbSNP rsID and a gene symbol search link (#6326)
 - Filter variants by OMIM inheritance pattern (#6355)
+- Gene panel filters to the Dark regions page (#6353)
 - More filter options for the gene variants page: function, inheritance pattern, predictors, frequency (#6357)
 ### Changed
 - Replaced Ensembl rest liftover service with liftover API from the Broad Institute (#6293)